wip(k8s): move transport to its file
